pyproject.toml
[tool.poetry]
name = "gdbundle-rtos"
version = "0.0.2"
description = ""
authors = ["Tyler Hoffman <tyler@memfault.com>"]
readme = "README.md"
include = ["gdbundle_rtos/scripts/*"]
license = "MIT"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Plugins",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 2",
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.4",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "*"
gdbundle = "~0.0.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
